[Refactor] Refactor List Protected Instances
